packages/shared-observability/pyproject.toml
[project]
name = "shared-observability"
version = "0.1.0"
description = "OpenTelemetry setup and structured logging"
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.28",
    "opentelemetry-sdk>=1.28",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "structlog>=24.4",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared_observability"]
